api sendet

+++ SmartPlantPot/Models/Messwerte.cs
@@ -8,6 +8,6 @@ namespace SmartPlantPot.Models
         public int Id { get; set; }
         public string Geraete_Id { get; set; }
         public DateTime Timestamp { get; set; }
-        public double Humidity { get; set; }
+        public decimal Humidity { get; set; }
     }
 }
